3813ictAssessmentChatApplication/src/app/app.component.spec.ts
import { TestBed, async, ComponentFixture } from '@angular/core/testing';
import { RouterTestingModule } from '@angular/router/testing';
import { AppComponent } from './app.component';
import { DebugElement } from '@angular/core';

import { FormsModule } from '@angular/forms';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { LoginComponent } from './login/login.component';

describe('AppComponent', () => {
  let app;
  let compiled;
  let fixture: ComponentFixture<AppComponent>;
  let dbe: DebugElement;
  beforeEach(async(() => {
    TestBed.configureTestingModule({
      imports: [
        RouterTestingModule.withRoutes(
          [{path: 'login', component: LoginComponent}]
        ),
        FormsModule,
        HttpClientTestingModule
      ],
      declarations: [
        AppComponent,
        LoginComponent
      ],
    }).compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(AppComponent);
    app = fixture.debugElement.componentInstance;
    compiled = fixture.debugElement.nativeElement;
    dbe = fixture.debugElement;
    fixture.detectChanges();
  });

  it('should create the app', () => {
    expect(app).toBeTruthy();
  });

  it('loginedIn should be false when first launch', () => {
    expect(app.loggedIn).toEqual(false);
  });

  // it('should render title in a h2 tag', () => {
  //   expect(compiled.querySelector('h2').textContent).toContain('Live Chat Application - Dashboard');
  // });
});
